+++ images/linux/scripts/installers/apache.sh
@@ -7,8 +7,10 @@
 # Install Apache
 apt-get install apache2 -y
 
-# Disable apache2.service
-systemctl is-active --quiet apache2.service && systemctl stop apache2.service
-systemctl disable apache2.service
+if [[ ! -f /run/systemd/container ]]; then
+    # Disable apache2.service
+    systemctl is-active --quiet apache2.service && systemctl stop apache2.service
+    systemctl disable apache2.service
+fi
 
 invoke_tests "WebServers" "Apache"
