+++ images/linux/scripts/installers/postgresql.sh
@@ -8,7 +8,7 @@ REPO_URL="https://apt.postgresql.org/pub/repos/apt/"
 
 #Preparing repo for PostgreSQL 12.
 wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
-echo "deb $REPO_URL `lsb_release -cs`-pgdg main" | tee /etc/apt/sources.list.d/pgdg.list
+echo "deb $REPO_URL $(lsb_release -cs)-pgdg main" | tee /etc/apt/sources.list.d/pgdg.list
 
 echo "Install PostgreSQL"
 apt update
@@ -17,12 +17,14 @@ apt install postgresql postgresql-client
 echo "Install libpq-dev"
 apt-get install libpq-dev
 
-# Disable postgresql.service
-systemctl is-active --quiet postgresql.service && systemctl stop postgresql.service
-systemctl disable postgresql.service
-
 rm /etc/apt/sources.list.d/pgdg.list
 
-echo "postgresql $REPO_URL" >> $HELPER_SCRIPTS/apt-sources.txt
+echo "postgresql $REPO_URL" >>"$HELPER_SCRIPTS"/apt-sources.txt
 
-invoke_tests "Databases" "PostgreSQL"
+if [[ ! -f /run/systemd/container ]]; then
+    # Disable postgresql.service
+    systemctl is-active --quiet postgresql.service && systemctl stop postgresql.service
+    systemctl disable postgresql.service
+
+    invoke_tests "Databases" "PostgreSQL"
+fi
